Fix several gcc warnings
Missing return in utils_convertMediaType
timeout used uninitialized in transaction_send
Various type mismatches between char * and uint8_t *
Return of void result in lwm2m_free
Mismatch int vs size_t in send_data fprintf.

+++ examples/shared/dtlsconnection.c
@@ -148,7 +148,7 @@ int send_data(dtls_connection_t *connP,
         port = saddr->sin6_port;
     }
 
-    fprintf(stderr, "Sending %d bytes to [%s]:%hu\r\n", length, s, ntohs(port));
+    fprintf(stderr, "Sending %d bytes to [%s]:%hu\r\n", (int)length, s, ntohs(port));
 
     output_buffer(stderr, buffer, length, 0);
 #endif
